feat: ✨ Merge page property & content changes
TYPO3 distinguishes between page properties and page content.
Therefore, there has never been a single field on a page that
stores the last editorial change, whether to properties or content.

Instead, TYPO3 stores changes to a page's properties
on the page itself and changes to content on a page
in the respective content elements.

However, editors expect both to belong together
when they ask “when was this page last updated?”.

To obtain a combined change date, these different values
must be aggregated via database queries.

With this change, a corresponding query is created
to identify “stale” pages for the editorial team.